% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_paf2gtf.R
\name{convert_gtf_granges}
\alias{convert_gtf_granges}
\title{convert gtf_tbl to granges object}
\usage{
convert_gtf_granges(gtf_tbl)
}
\arguments{
\item{gtf_tbl}{A gtf_oriented tibble object.}
}
\value{
A granges object.
}
\description{
convert gtf_tbl to granges object
}
