% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_paf.R
\name{convert_paf_variant}
\alias{convert_paf_variant}
\title{convert paf data into a tibble suitable for "isodata" class.}
\usage{
convert_paf_variant(paf, sample_name, withcare = F, verbose = T,
  parallel = T)
}
\arguments{
\item{paf}{A tbl. A paf file generated by minimap2 with --cs argument and read by read_paf.}

\item{sample_name}{A one-length character vector. Used to fill the "sample" column.}

\item{withcare}{Logical. If true, stop when a transcript's end position is NOT consistent between described one and calculated one.}

\item{verbose}{Logical.}

\item{parallel}{Logical. If true, use furrr::future_map instead of purrr::map}
}
\value{
A tibble with columns: seq_id, chr, type, start, end, ref, alt, and sample.
start, end: 0-based. the bases between them present in ref. If insertion, start == end. If point mutation, start + 1 == end.
}
\description{
"type" column contains "+", "-", "*", or "~".
"+": insertion, "-": deletion, "*": substitution, "~": spliced intron.
These indicators are equal to those used in minimap2.
}
