% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_sam}
\alias{export_sam}
\title{export isodata as SAM format.}
\usage{
export_sam(isodata, path, sample_name, ref_genome)
}
\arguments{
\item{isodata}{An isodata class object.}

\item{path}{A file path to create.}

\item{sample_name}{A sample name to export.}

\item{ref_genome}{A BSgenome class object. It is used to construct SEQ segment and SQ tags.}
}
\description{
You MUST confirm that all variants lie ONLY on exons.
Some arguments are set to fixed values:
MAPQ = 255, RNEXT = asterisk, PNEXT = 0, TLEN = 0, QUAL = asterisk.
}
