% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant.R
\name{filter_variant}
\alias{filter_variant}
\title{Filtering 'variant' slot using 'exon_ranges' slot and provided true variant tibble.}
\usage{
filter_variant(variant, true_variant, ref_genome)
}
\arguments{
\item{variant}{A tibble}

\item{true_variant}{A tibble containing variant information to be retained. It is usually generated from WES or WGS data.}

\item{ref_genome}{A BSgenome class object.}
}
\description{
Filtering 'variant' slot using 'exon_ranges' slot and provided true variant tibble.
}
