% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant.R
\name{filter_variant_isodata}
\alias{filter_variant_isodata}
\title{Filtering 'variant' slot using 'exon_ranges' slot and provided true variant tibble.}
\usage{
filter_variant_isodata(isodata, df = NULL, ref_genome = NULL,
  .restrict_on_exon = T)
}
\arguments{
\item{isodata}{An isodata class object with 'variant' and 'exon_ranges' slots.}

\item{df}{A tibble containing variant information to be retained. It is usually generated from WES or WGS data.
If unspecified, filtering is conducted using only 'exon_ranges' slot.}

\item{ref_genome}{A BSgenome class object. if specified, spliced introns that exist on exons in exon_range slot will be treated as deletion.
If unspecified, spliced introns will be ignored.}

\item{.restrict_on_exon}{A logical. Whether filtering variants to retain only those on exons of correlated transcripts.}
}
\description{
Filtering 'variant' slot using 'exon_ranges' slot and provided true variant tibble.
}
