% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{iso_append}
\alias{iso_append}
\title{append data to an isodata class object.}
\usage{
iso_append(isodata, obj, append = T)
}
\arguments{
\item{isodata}{An isodata class object.}

\item{obj}{A data to be imported.}

\item{append}{Logical. If false, old data in the slot concerned will be deleted.}
}
\description{
"isodata" class object consists of many slots of tibbles.
It is aimed to be able to store almost all information about full-length transcripts.
"import_iso" is a S4 generic for importing data into an isodata class object.
}
