% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isocan.R
\docType{package}
\name{isocan}
\alias{isocan}
\alias{isocan-package}
\title{IsoCan: A package for handling long-read isoforms in cancer.}
\description{
The IsoCan package provides *** categories of analytic functions:
***, ***, and ***.
}
