% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{isodata-class}
\alias{isodata-class}
\alias{isodata}
\title{S4 class for data storage}
\description{
You should NOT use this function for inputting your data because data stored in different slots may be inconsistent with each other.
Instead, you can use 'store_isodata' function to input and extend your data,
and next you can use 'construct_isodata' for consistent position data storage in exon, transcript, and gene levels.
For checking if your isodata object has consistent structure, use 'check_isodata_consistent'.
}
\section{Slots}{

\describe{
\item{\code{exon_ranges}}{A tibble for exon-level information, including exons' positions and additional meta data.
Any positional data should be stored first in this slot, and higher-level positional data should be constructed by the function 'construct_isodata'.}

\item{\code{gene_attr}}{A tibble for gene-level information, including transcripts' positions and additional meta data.}

\item{\code{tx_attr}}{A tibble for transcript-level information, possibly including transcripts' positions and additional meta data.}

\item{\code{variant}}{A tibble for variant information.}

\item{\code{tx_seq_correspondence}}{A tibble with 'tx_id' and 'seq_id' columns.}

\item{\code{gene_exp}}{A tibble for gene-level expression.}

\item{\code{tx_exp}}{A tibble for transcript-level expression.}

\item{\code{exp_type}}{1-length character vector. For the time being, only "TPM" is allowed.}

\item{\code{sample_group}}{A character vector. It must have the smaller length than colnames of tx_exp by one.}
}}

