% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isodata_append.R
\name{isodata_append.GRanges}
\alias{isodata_append.GRanges}
\title{import a GRanges object into an isodata class object.}
\usage{
\method{isodata_append}{GRanges}(obj, isodata, append)
}
\arguments{
\item{obj}{A data to be imported.}

\item{isodata}{An isodata class object.}

\item{append}{Logical. If false, old data in the slot concerned will be deleted.}
}
\description{
import a GRanges object into an isodata class object.
}
