% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_paf.R
\name{read_paf}
\alias{read_paf}
\title{read a *.paf file}
\usage{
read_paf(file, .query_position = FALSE)
}
\arguments{
\item{file}{A one-length char vector.}

\item{.query_position}{A logical.}
}
\value{
a tibble with columns: seq_id, query_length, strand, chr, start, end, and cs.
}
\description{
read a paf file generated by minimap2 with --cs argument.
It is a wrapper of readr::read_tsv.
}
