% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_paf2gtf.R
\name{retrieve_gtf}
\alias{retrieve_gtf}
\title{retreive gtf_oriented tibble from paf}
\usage{
retrieve_gtf(paf, sample_name, retreive_type = "exon", withcare = F,
  verbose = T, parallel = T, paf_variants = NULL)
}
\arguments{
\item{paf}{A tbl. A paf file generated by minimap2 with --cs argument and read by read_paf.}

\item{sample_name}{A one-length character vector. Used to fill the "sample" column.}

\item{retreive_type}{One of 'exon', 'intron', 'tx'(='transcript').}

\item{withcare}{Logical. If true, stop when a transcript's end position is NOT consistent between described one and calculated one.}

\item{verbose}{Logical.}

\item{parallel}{Logical. If true, use furrr::future_map instead of purrr::map}

\item{paf_variants}{A tibble generated by convert_paf_variant function. If specified, common procedure with convert_paf_variant will be skipped.}
}
\value{
A gtf_oriented tibble object. 1-based coordination.
}
\description{
Retreive granges object from paf data and paf-based variant data
}
